requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-classy-tags==0.8.0
django-disqus==0.5
django-taggit==0.22.2
django-taggit-templatetags2==1.6.1
idna==2.7
Markdown==2.6.11
martor==1.2.7
mock==2.0.0
pbr==4.2.0
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
